extract util used by jsonmergepatch and SMPatch
